Fix TypeError in tsc_utils.py related to type hinting
This pull request addresses the TypeError we were facing when trying to use indexing or subscription on an object that doesn't support it. The problem was due to the way we were type hinting dictionaries before Python 3.9.

Changes:
Imported necessary classes (Dict and List) from the typing module.
Updated the type hint for last_helds to use Dict and List.